pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "relaylm"
version = "0.1.0"
description = "Single-command AI router environment bootstrap"
requires-python = ">=3.11"
license = "MIT"
authors = [
    {name = "RelayLM Contributors"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",

    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Distributed Computing",
]
dependencies = [
    "typer>=0.12",
    "pyyaml>=6.0",
    "keyring>=25.0",
    "huggingface-hub>=0.24",
    "pydantic>=2.0",
]

[project.urls]
Homepage = "https://github.com/anomalyco/relaylm"
Repository = "https://github.com/anomalyco/relaylm"
Issues = "https://github.com/anomalyco/relaylm/issues"

[project.scripts]
relaylm = "relaylm.cli.app:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 88
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
files = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
